+++ games/tictactoe/__init__.py
@@ -1,0 +1,2 @@
+from .tictactoe import *
+from .helperfunctions import *
