--- games/tictactoe/exceptions.py
@@ -1,14 +1,0 @@
-class GameEx(Exception):
-    pass
-
-
-class GuessEx(GameEx):
-    pass
-
-
-class GameOverEx(GameEx):
-    pass
-
-
-class WinEx(GameOverEx):
-    pass
